spark/requirements.txt
# requirements.txt
pandas==1.2.5
pendulum==2.0.5
boto3==1.20.54
tabulate==0.9.0
requests==2.31.0
urllib3==1.26.16
pyarrow==9.0.0
pycryptodome==3.15.0